Minimize CLI dependencies
