+++ hydro_cli/Cargo.toml
@@ -28,6 +28,6 @@ serde = { version = "1", features = [ "derive" ] }
 serde_json = "1"
 tempfile = "3.3.0"
 async-ssh2-lite = { version = "0.4.2", features = [ "tokio", "vendored-openssl" ] }
-hydroflow = { path = "../hydroflow", features = ["cli_integration"] }
+hydroflow_cli_integration = { path = "../hydroflow_cli_integration" }
 
 [dev-dependencies]
