+++ hydroflow/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 default = [ "async", "macros" ]
 async = [ "futures" ]
 macros = [ "hydroflow_macro", "hydroflow_datalog" ]
-cli_integration = [ "tempfile", "async-trait" ]
+cli_integration = [ "hydroflow_cli_integration", "async-trait" ]
 
 [[example]]
 name = "echoserver"
@@ -69,7 +69,7 @@ static_assertions = "1.1.0"
 tokio-stream = { version = "0.1.10", features = [ "io-util" ] }
 variadics = { path = "../variadics" }
 rustc-hash = "1.1.0"
-tempfile = { version = "3.3.0", optional = true }
+hydroflow_cli_integration = { optional = true, path = "../hydroflow_cli_integration", features = [ "runtime" ] }
 async-trait = { version = "0.1", optional = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
