+++ hydroflow_cli_integration/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "hydroflow_cli_integration"
+version = "0.1.0"
+edition = "2021"
+
+[features]
+default = []
+runtime = [ "tempfile", "async-trait", "bytes", "tokio", "futures" ]
+
+[dependencies]
+serde = { version = "1", features = [ "derive" ] }
+serde_json = "1"
+tempfile = { version = "3.3.0", optional = true }
+async-trait = { version = "0.1", optional = true }
+bytes = { version = "1.1.0", optional = true }
+futures = { version = "0.3.26", optional = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+tokio = { version = "1.16", features = [ "full" ], optional = true }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+tokio = { version = "1.16", features = [ "rt" , "sync", "macros", "io-util", "time" ], optional = true }
